# Strengthen index simplification for cast epilogue matmul