… tests/cpp/test_expr_simplifier.cpp
… tests/cpp/test_loop_rotation.cpp