@jacobhinkle 2024-02-27
Changing the size in `NVFuserTest.FusionAmpereMatmulSmemEpilogueCast_CUDA` to 16384, 16384, 256 (i.e. a large memory-bound problem size), we go from 398 GB/s to 554 GB/s on A100 80GB PCIe. That's a 39% speedup! That roughly matches our internal measurements that output bandwidth for fp16 outputs is about half that of fp32 outputs. Seems like it's worth figuring out the proof slowdown here since this optimization can have a large impact when output writes are a significant portion of runtime.