@zasdfgbnm 2024-02-27
> That's a 39% speedup!

That's a lot! Thanks for measuring this!

> I also measured after manually performing the last optimization (`i134 % 8 -> i134`, `i134 / 8 -> 0`, hoisting) and saw now effect. It's possible these are already performed by the cuda compiler.

Even if there is no visible perf improvement, I still suggest go ahead and implementing it because:
1. It should be easy to implement
2. It improves code readability