@jacobhinkle 2024-03-01
Summary: As of now, this PR causes a 27% slowdown in total runtime for the test mentioned in the PR description compared to `main`. That is an improvement over my original method that showed 2.2x runtime due to a recently pushed change. In debug mode we currently see 3.8x runtime, up from 14x.

The latest pushed change switches from unordered_set method to just limiting the recursion depth directly to 2.

Release mode test timing:
```
recursion depth  time (sec)  Simplified?
      0              1.5         no
      1              1.5         no
      2              1.9         yes
      3              9.3         yes
      4            133           yes
Compare to
 main                1.5         no
 unordered_set       3.3         yes
```
The Release compilation mode seems to make a big difference.

Debug build timings:
```
recursion depth  time (sec)  Simplified?
      0              7.4         no
      1              7.5         no
      2             28           yes
      3            422           yes
      4              ?           yes
 main                7.4         no
 unordered_set     104           yes
```