@jacobhinkle 2024-03-01
I'm now trying to write a good test for this PR, then I will clean up and mark it ready.