@jacobhinkle 2024-03-12
Recently-pushed change strengthens `eliminateTrivialComputation` to simplify `a % b` and `a / b` when `-|b| < a < |b|`. This leads to a very nice kernel:
```c++
    // main loop
  }
  __syncthreads();
  #pragma unroll
  for(nvfuser_index_t i117 = 0; i117 < 4; ++i117) {
    nvfuser_index_t i118;
    i118 = 32 * i117;
    nvfuser_index_t i119;
    i119 = i53 + (2048LL * i117);
    #pragma unroll
    for(nvfuser_index_t i120 = 0; i120 < 8; ++i120) {
      nvfuser_index_t i121;
      i121 = i118 + (4 * i120);
      nvfuser_index_t i122;
      i122 = i30 + i120;
      nvfuser_index_t i123;
      i123 = (i119 + (32LL * (i122 / 4))) + (8LL * (i54 ^ (i122 % 4)));
      #pragma unroll
      for(nvfuser_index_t i124 = 0; i124 < 2; ++i124) {
        loadGeneric<float, 2>( &T8[(i123 + (1024LL * i124))],  &T3[(i121 + (2LL * i124))]);
      }
    }
  }
  __syncthreads();
  #pragma unroll
  for(nvfuser_index_t i125 = 0; i125 < 16; ++i125) {
    nvfuser_index_t i126;
    i126 = i55 + (1024 * i125);
    Array<__half, 8, 8> T7;
    #pragma unroll
    for(nvfuser_index_t i127 = 0; i127 < 8; ++i127) {
      T7[i127]
         = __float2half(T8[(i126 + i127)]);
    }
    if ((b58 && (i67 < (-(8 * i125))))) {
      loadLocalToGlobal<__half, /*vec_size=*/8, /*is_volatile=*/false>( &T4[(i57 + (i10 * i125))], &T7[0]);
    }
  }
}
```
However, it feels a bit slower. I'm going to evaluate the compile time with and without this change.