@jacobhinkle 2024-03-20
> it feels a bit slower.

On my machine this takes overall test time from 1.9 to 3.9 seconds in a release build and from 7.4 to 15 seconds in debug build. I think it's probably best to leave this new optimization mentioned in the last comment for another PR so that I can play more with speeding it up. For now, I will revert it and ensure the timing makes sense then hopefully we can merge.