@jacobhinkle 2024-03-20
With reverted commit, compile times are back down. If tests pass, I think this PR is ready.