@jacobhinkle 2024-03-21
After #1972 the compile time is down to 1.3s. Even better, since it re-used lots of proofs, re-enabling the reverted simplification actually doesn't appreciably change compile time!